Add measure-by-weight option, friendlier inventory labels with tap-for-help, and test suite
- Inventory items can now be tracked by weight/volume (oz, lb, g, kg, fl oz, ml, L)
  instead of count; stock, prices, alerts, and sales all become per-unit and accept
  fractional amounts (e.g. 90.5 oz of candle wax)
- Renamed confusing fields based on tester feedback (Cost/unit -> Your cost,
  Sale price -> Selling price, Quantity on hand -> How many in stock,
  Min. stock level -> Low stock alert) and added tappable ? help icons with
  plain-language explanations that adapt to the chosen unit
- Quantities are decimal end-to-end; existing databases migrate automatically
- Fixed BOM deductions being rounded to whole numbers (2.5 oz deducted as 3)
- BOM component deductions now always follow a sale, going negative rather than
  silently skipping when stock runs short; editing a sale now cascades the
  quantity change to BOM components too
- Added 33 tests: BOM cascade, adjustments, low-stock alerts, selling by weight,
  sale-edit deltas, and real-file database migration coverage; fixed test project
  not compiling since the BOM feature
- Added GitHub Actions workflow to run unit tests on every push

+++ StockAndFlow.Core/Models/BusinessMetrics.cs
@@ -14,7 +14,7 @@ namespace StockAndFlow.Models
         public decimal OverallProfitMargin { get; set; }
 
         public int TotalSalesCount { get; set; }
-        public int TotalItemsSold { get; set; }
+        public decimal TotalItemsSold { get; set; }
         public int UniqueInventoryItems { get; set; }
 
         public DateTime CalculatedAt { get; set; }
